1821: Fix issue with possible log injection entry point

+++ domain-ee/ee-max-cfi-app/src/python_src/pydantic_models.py
@@ -7,7 +7,7 @@ class MaxRatingsForClaimForIncreaseRequest(BaseModel):
 
 class Rating(BaseModel):
     diagnostic_code: int
-    max_rating: float
+    max_rating: int
 
 
 class MaxRatingsForClaimForIncreaseResponse(BaseModel):
